requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
attrs==26.1.0
frozenlist==1.8.0
idna==3.12
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
propcache==0.4.1
Pygments==2.20.0
python-dotenv==1.2.2
rich==15.0.0
yarl==1.23.0
